requirements.txt
python-binance==1.0.16
pandas==1.4.2
numpy==1.22.4
nest-asyncio==1.5.5
mplfinance==0.12.9b1
